fix clippy warnings

+++ tests/src/utils/common.rs
@@ -27,11 +27,11 @@ pub fn signed_int_to_float(amount: i128, decimals: i32) -> f64 {
 }
 
 pub fn float_to_int(amount: f64, decimals: u32) -> u128 {
-    (amount as f64 * 10.0f64.powi(decimals as i32)) as u128
+    (amount * 10.0f64.powi(decimals as i32)) as u128
 }
 
 pub fn int_to_float(amount: u128, decimals: u32) -> f64 {
-    ((amount as f64) / 10.0f64.powi(decimals as i32)) as f64
+    (amount as f64) / 10.0f64.powi(decimals as i32)
 }
 
 pub fn float_to_int_sp(amount: f64) -> u128 {
